requirements.txt
# Obviously you might want specific versions here.
torch
polars
lightning
pyyaml
# Optional, but I like it
black
tensorboard
# Maybe
logzero
mlflow
pybaseball==2.2.7
morphers==0.1.2
